// src/zap/schema.ts
export class Field_Slot extends $.Struct {
  static readonly _zap = {
    displayName: "slot",
    id: "c42305476bb4746f",
    size: new $.ObjectSize(24, 4),
  };
  /**
* Offset, in units of the field's size, from the beginning of the section in which the field
* resides.  E.g. for a UInt32 field, multiply this by 4 to get the byte offset from the
* beginning of the data section.
*
*/
  get offset(): number {
    return $.utils.getUint32(4, this);
  }
  set offset(value: number) {
    $.utils.setUint32(4, value, this);
  }
  _adoptType(value: $.Orphan<Type>): void {
    $.utils.adopt(value, $.utils.getPointer(2, this));
  }
  _disownType(): $.Orphan<Type> {
    return $.utils.disown(this.type);
  }
  get type(): Type {
    return $.utils.getStruct(2, Type, this);
  }
  _hasType(): boolean {
    return !$.utils.isNull($.utils.getPointer(2, this));
  }
  _initType(): Type {
    return $.utils.initStructAt(2, Type, this);
  }
  set type(value: Type) {
    $.utils.copyFrom(value, $.utils.getPointer(2, this));
  }
  _adoptDefaultValue(value: $.Orphan<Value>): void {
    $.utils.adopt(value, $.utils.getPointer(3, this));
  }
  _disownDefaultValue(): $.Orphan<Value> {
    return $.utils.disown(this.defaultValue);
  }
  get defaultValue(): Value {
    return $.utils.getStruct(3, Value, this);
  }
  _hasDefaultValue(): boolean {
    return !$.utils.isNull($.utils.getPointer(3, this));
  }
  _initDefaultValue(): Value {
    return $.utils.initStructAt(3, Value, this);
  }
  set defaultValue(value: Value) {
    $.utils.copyFrom(value, $.utils.getPointer(3, this));
  }
  /**
* Whether the default value was specified explicitly.  Non-explicit default values are always
* zero or empty values.  Usually, whether the default value was explicit shouldn't matter.
* The main use case for this flag is for structs representing method parameters:
* explicitly-defaulted parameters may be allowed to be omitted when calling the method.
*
*/
  get hadExplicitDefault(): boolean {
    return $.utils.getBit(128, this);
  }
  set hadExplicitDefault(value: boolean) {
    $.utils.setBit(128, value, this);
  }
  toString(): string { return "Field_Slot_" + super.toString(); }
}
/**
* A group.
*
*/
export class Field_Group extends $.Struct {
  static readonly _zap = {
    displayName: "group",
    id: "cafccddb68db1d11",
    size: new $.ObjectSize(24, 4),
  };
  /**
* The ID of the group's node.
*
*/
  get typeId(): bigint {
    return $.utils.getUint64(16, this);
  }
  set typeId(value: bigint) {
    $.utils.setUint64(16, value, this);
  }
  toString(): string { return "Field_Group_" + super.toString(); }
}
export const Field_Ordinal_Which = {
  IMPLICIT: 0,
  /**
* The original ordinal number given to the field.  You probably should NOT use this; if you need
* a numeric identifier for a field, use its position within the field array for its scope.
* The ordinal is given here mainly just so that the original schema text can be reproduced given
* the compiled version -- i.e. so that `zap compile -ozap` can do its job.
*
*/
  EXPLICIT: 1
} as const;
export type Field_Ordinal_Which = (typeof Field_Ordinal_Which)[keyof typeof Field_Ordinal_Which];
export class Field_Ordinal extends $.Struct {
  static readonly IMPLICIT = Field_Ordinal_Which.IMPLICIT;
  static readonly EXPLICIT = Field_Ordinal_Which.EXPLICIT;
  static readonly _zap = {
    displayName: "ordinal",
    id: "bb90d5c287870be6",
    size: new $.ObjectSize(24, 4),
  };
  get _isImplicit(): boolean {
    return $.utils.getUint16(10, this) === 0;
  }
  set implicit(_: true) {
    $.utils.setUint16(10, 0, this);
  }
  /**
* The original ordinal number given to the field.  You probably should NOT use this; if you need
* a numeric identifier for a field, use its position within the field array for its scope.
* The ordinal is given here mainly just so that the original schema text can be reproduced given
* the compiled version -- i.e. so that `zap compile -ozap` can do its job.
*
*/
  get explicit(): number {
    $.utils.testWhich("explicit", $.utils.getUint16(10, this), 1, this);
    return $.utils.getUint16(12, this);
  }
  get _isExplicit(): boolean {
    return $.utils.getUint16(10, this) === 1;
  }
  set explicit(value: number) {
    $.utils.setUint16(10, 1, this);
    $.utils.setUint16(12, value, this);
  }
  toString(): string { return "Field_Ordinal_" + super.toString(); }
  which(): Field_Ordinal_Which {
    return $.utils.getUint16(10, this) as Field_Ordinal_Which;
  }
}
export const Field_Which = {
  SLOT: 0,
  /**
* Indicates where this member appeared in the code, relative to other members.
* Code ordering may have semantic relevance -- programmers tend to place related fields
* together.  So, using code ordering makes sense in human-readable formats where ordering is
* otherwise irrelevant, like JSON.  The values of codeOrder are tightly-packed, so the maximum
* value is count(members) - 1.  Fields that are members of a union are only ordered relative to
* the other members of that union, so the maximum value there is count(union.members).
*
*/
  GROUP: 1
} as const;
export type Field_Which = (typeof Field_Which)[keyof typeof Field_Which];
/**
* Schema for a field of a struct.
*
*/
export class Field extends $.Struct {
  static readonly NO_DISCRIMINANT = 65535;
  static readonly SLOT = Field_Which.SLOT;
  static readonly GROUP = Field_Which.GROUP;
  static readonly _zap = {
    displayName: "Field",
    id: "9aad50a41f4af45f",
    size: new $.ObjectSize(24, 4),
    defaultDiscriminantValue: $.getUint16Mask(65535)
  };
  static _Annotations: $.ListCtor<Annotation>;
  get name(): string {
    return $.utils.getText(0, this);
  }
  set name(value: string) {
    $.utils.setText(0, value, this);
  }
  /**
* Indicates where this member appeared in the code, relative to other members.
* Code ordering may have semantic relevance -- programmers tend to place related fields
* together.  So, using code ordering makes sense in human-readable formats where ordering is
* otherwise irrelevant, like JSON.  The values of codeOrder are tightly-packed, so the maximum
* value is count(members) - 1.  Fields that are members of a union are only ordered relative to
* the other members of that union, so the maximum value there is count(union.members).
*
*/
  get codeOrder(): number {
    return $.utils.getUint16(0, this);
  }
  set codeOrder(value: number) {
    $.utils.setUint16(0, value, this);
  }
  _adoptAnnotations(value: $.Orphan<$.List<Annotation>>): void {
    $.utils.adopt(value, $.utils.getPointer(1, this));
  }
  _disownAnnotations(): $.Orphan<$.List<Annotation>> {
    return $.utils.disown(this.annotations);
  }
  get annotations(): $.List<Annotation> {
    return $.utils.getList(1, Field._Annotations, this);
  }
  _hasAnnotations(): boolean {
    return !$.utils.isNull($.utils.getPointer(1, this));
  }
  _initAnnotations(length: number): $.List<Annotation> {
    return $.utils.initList(1, Field._Annotations, length, this);
  }
  set annotations(value: $.List<Annotation>) {
    $.utils.copyFrom(value, $.utils.getPointer(1, this));
  }
  /**
* If the field is in a union, this is the value which the union's discriminant should take when
* the field is active.  If the field is not in a union, this is 0xffff.
*
*/
  get discriminantValue(): number {
    return $.utils.getUint16(2, this, Field._zap.defaultDiscriminantValue);
  }
  set discriminantValue(value: number) {
    $.utils.setUint16(2, value, this, Field._zap.defaultDiscriminantValue);
  }
  /**
* A regular, non-group, non-fixed-list field.
*
*/
  get slot(): Field_Slot {
    $.utils.testWhich("slot", $.utils.getUint16(8, this), 0, this);
    return $.utils.getAs(Field_Slot, this);
  }
  _initSlot(): Field_Slot {
    $.utils.setUint16(8, 0, this);
    return $.utils.getAs(Field_Slot, this);
  }
  get _isSlot(): boolean {
    return $.utils.getUint16(8, this) === 0;
  }
  set slot(_: true) {
    $.utils.setUint16(8, 0, this);
  }
  /**
* A group.
*
*/
  get group(): Field_Group {
    $.utils.testWhich("group", $.utils.getUint16(8, this), 1, this);
    return $.utils.getAs(Field_Group, this);
  }
  _initGroup(): Field_Group {
    $.utils.setUint16(8, 1, this);
    return $.utils.getAs(Field_Group, this);
  }
  get _isGroup(): boolean {
    return $.utils.getUint16(8, this) === 1;
  }
  set group(_: true) {
    $.utils.setUint16(8, 1, this);
  }
  get ordinal(): Field_Ordinal {
    return $.utils.getAs(Field_Ordinal, this);
  }
  _initOrdinal(): Field_Ordinal {
    return $.utils.getAs(Field_Ordinal, this);
  }
  toString(): string { return "Field_" + super.toString(); }
  which(): Field_Which {
    return $.utils.getUint16(8, this) as Field_Which;
  }
}
// ...
export const Type_Which = {
  VOID: 0,
  BOOL: 1,
  INT8: 2,
  INT16: 3,
  INT32: 4,
  INT64: 5,
  UINT8: 6,
  UINT16: 7,
  UINT32: 8,
  UINT64: 9,
  FLOAT32: 10,
  FLOAT64: 11,
  TEXT: 12,
  DATA: 13,
  LIST: 14,
  ENUM: 15,
  STRUCT: 16,
  INTERFACE: 17,
  ANY_POINTER: 18
} as const;
export type Type_Which = (typeof Type_Which)[keyof typeof Type_Which];
// ...
export class Brand extends $.Struct {
  static readonly Scope = Brand_Scope;
  static readonly Binding = Brand_Binding;
  static readonly _zap = {
    displayName: "Brand",
    id: "903455f06065422b",
    size: new $.ObjectSize(0, 1),
  };
  static _Scopes: $.ListCtor<Brand_Scope>;
  _adoptScopes(value: $.Orphan<$.List<Brand_Scope>>): void {
    $.utils.adopt(value, $.utils.getPointer(0, this));
  }
  _disownScopes(): $.Orphan<$.List<Brand_Scope>> {
    return $.utils.disown(this.scopes);
  }
  /**
* For each of the target type and each of its parent scopes, a parameterization may be included
* in this list. If no parameterization is included for a particular relevant scope, then either
* that scope has no parameters or all parameters should be considered to be `AnyPointer`.
*
*/
  get scopes(): $.List<Brand_Scope> {
    return $.utils.getList(0, Brand._Scopes, this);
  }
  _hasScopes(): boolean {
    return !$.utils.isNull($.utils.getPointer(0, this));
  }
  _initScopes(length: number): $.List<Brand_Scope> {
    return $.utils.initList(0, Brand._Scopes, length, this);
  }
  set scopes(value: $.List<Brand_Scope>) {
    $.utils.copyFrom(value, $.utils.getPointer(0, this));
  }
  toString(): string { return "Brand_" + super.toString(); }
}
// ...
export class Annotation extends $.Struct {
  static readonly _zap = {
    displayName: "Annotation",
    id: "f1c8950dab257542",
    size: new $.ObjectSize(8, 2),
  };
  /**
* ID of the annotation node.
*
*/
  get id(): bigint {
    return $.utils.getUint64(0, this);
  }
  set id(value: bigint) {
    $.utils.setUint64(0, value, this);
  }
  _adoptBrand(value: $.Orphan<Brand>): void {
    $.utils.adopt(value, $.utils.getPointer(1, this));
  }
  _disownBrand(): $.Orphan<Brand> {
    return $.utils.disown(this.brand);
  }
  /**
* Brand of the annotation.
*
* Note that the annotation itself is not allowed to be parameterized, but its scope might be.
*
*/
  get brand(): Brand {
    return $.utils.getStruct(1, Brand, this);
  }
  _hasBrand(): boolean {
    return !$.utils.isNull($.utils.getPointer(1, this));
  }
  _initBrand(): Brand {
    return $.utils.initStructAt(1, Brand, this);
  }
  set brand(value: Brand) {
    $.utils.copyFrom(value, $.utils.getPointer(1, this));
  }
  _adoptValue(value: $.Orphan<Value>): void {
    $.utils.adopt(value, $.utils.getPointer(0, this));
  }
  _disownValue(): $.Orphan<Value> {
    return $.utils.disown(this.value);
  }
  get value(): Value {
    return $.utils.getStruct(0, Value, this);
  }
  _hasValue(): boolean {
    return !$.utils.isNull($.utils.getPointer(0, this));
  }
  _initValue(): Value {
    return $.utils.initStructAt(0, Value, this);
  }
  set value(value: Value) {
    $.utils.copyFrom(value, $.utils.getPointer(0, this));
  }
  toString(): string { return "Annotation_" + super.toString(); }
}